Feature/huiswerkbegeleiding (#51)
* Added homework assistance to front-and backend

* Added selenium test

* updated homework entries

* styling

* fixed selenium test

* changed the entry form, its now connected to the activities

* activity 1 not editable because its the homework activity

* Selenium fix

* selenium fix

+++ resources/views/admin/activity/index.blade.php
@@ -43,11 +43,13 @@
                            title="Edit"><i class="fa fa-pencil" aria-hidden="true"></i></a>
                     </td>
                     <td>
-                        {{ Form::open(['id' => 'formDelete'.$activity->id,'method' => 'DELETE', 'route' => ['activiteit.destroy', $activity->id]]) }}
-                            <a title="Delete" href="javascript:void(0)" onclick="document.getElementById('formDelete{{$activity->id}}').submit()">
-                                <i class="fa fa-times" aria-hidden="true"></i>
-                            </a>
-                        {{ Form::close() }}
+                        @if ($activity->id !== 1)
+                            {{ Form::open(['id' => 'formDelete'.$activity->id,'method' => 'DELETE', 'route' => ['activiteit.destroy', $activity->id]]) }}
+                                <a title="Delete" href="javascript:void(0)" onclick="document.getElementById('formDelete{{$activity->id}}').submit()">
+                                    <i class="fa fa-times" aria-hidden="true"></i>
+                                </a>
+                            {{ Form::close() }}
+                        @endif
                     </td>
                 </tr>
             @endforeach
